added fix for linting issue
